update_04.05.2021

+++ int-dmp-core/src/main/java/test/intdmp/core/service/repository/FileSystemRepository.java
@@ -1,7 +1,14 @@
 package test.intdmp.core.service.repository;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+import test.intdmp.core.configuration.FileDirectoryConfiguration;
+import test.intdmp.core.model.files._enum.UploadStatus;
+import test.intdmp.core.service.fileManager._class.FileExistsException;
 
+import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -9,16 +16,23 @@ import java.nio.file.Paths;
 @Repository
 public class FileSystemRepository {
 
-    String resources_dir = FileSystemRepository.class.getResource("/")
-            .getPath();
+    @Autowired
+    private FileDirectoryConfiguration fileDirectoryConfiguration;
 
-    public String save(byte[] content, String name) throws Exception {
-        Path newFile = Paths.get(resources_dir + name);
+
+    public FileSystemRepository() throws URISyntaxException {
+    }
+
+    public String save(byte[] content, String name) throws IOException {
+        Path newFile = Paths.get(fileDirectoryConfiguration.saveDirectory() + "\\" + name);
         Files.createDirectories(newFile.getParent());
 
-        Files.write(newFile, content);
-
-        return newFile.toAbsolutePath().toString();
+        if (Files.notExists(newFile)) {
+            Files.write(newFile, content);
+            return newFile.toAbsolutePath().toString();
+        }
+        else
+            throw new FileExistsException(UploadStatus.FILE_EXIST);
     }
 
 }
